Fix: deprecated array access syntax on PHP
The "Array and string offset access syntax with curly braces is
deprecated" error appears on PHP7_4 and above.

This error is causing the behat and PHPUnit tests of the MathType Moodle
plugin suite to fail both locally and on the TravisCI pipeline.

This commit includes a simple change that fixes the issue by using a
more up-to-date syntax to acess an array.

We're planning to publish a new release of the MathType Moodle plugin
suite. The code of this version uses the same branch name as this one:
'release-7.25', on the affected projects on github: filter, atto and
tinymce.